# docs/swagger.yaml
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A web service that, upon request, collects data from the public API
    of the national bank and saves the data to the local TEST database
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger kursRates API
  version: "0.1"
paths:
  /currency/{date}:
    get:
      consumes:
      - application/json
      - application/json
      description: |-
        Get currency data for a specific date.
        Get currency data for a specific date and currency code.
      parameters:
      - description: Date in DD.MM.YYYY format
        in: path
        name: date
        required: true
        type: string
      responses: {}
      summary: Get currency data by date and code
      tags:
      - currency
      - currency
  /currency/{date}/{code}:
    get:
      consumes:
      - application/json
      - application/json
      description: |-
        Get currency data for a specific date.
        Get currency data for a specific date and currency code.
      parameters:
      - description: Date in DD.MM.YYYY format
        in: path
        name: date
        required: true
        type: string
      - description: Currency code (e.g., USD)
        in: path
        name: code
        required: true
        type: string
      responses: {}
      summary: Get currency data by date and code
      tags:
      - currency
      - currency
  /currency/save/{date}:
    post:
      consumes:
      - application/json
      description: Save currency data for a specific date.
      parameters:
      - description: Date in DD.MM.YYYY format
        in: path
        name: date
        required: true
        type: string
      responses: {}
      summary: Save currency data
      tags:
      - currency
  /health:
    get:
      description: Returns the health status of the application, including the database
        availability.
      operationId: health-check
      produces:
      - application/json
      responses:
        "200":
          description: 'Status: Available'
          schema:
            type: string
        "503":
          description: 'Status: Not available'
          schema:
            type: string
      summary: Check the health status of the application
swagger: "2.0"